Resolved:NullPointerException at Line96 Incident_guardfragment
Moved Retrofit to IncidentAdapter
