+++ app/src/main/java/com/example/vmsandroid/IncidentAdapter.java
@@ -15,14 +15,23 @@ import androidx.recyclerview.widget.RecyclerView;
 import java.util.ArrayList;
 import java.util.HashMap;
 
+import okhttp3.OkHttpClient;
+import okhttp3.logging.HttpLoggingInterceptor;
 import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
+import retrofit2.Retrofit;
+import retrofit2.converter.gson.GsonConverterFactory;
+
 
 public class IncidentAdapter extends RecyclerView.Adapter<IncidentAdapter.IncidentViewHolder> {
     private ArrayList<IterateIncident> mIncidentList;
 
     public static class IncidentViewHolder extends RecyclerView.ViewHolder {
+        private Retrofit retrofit;
+        private RetrofitInterface retrofitInterface;
+        private String BASE_URL = "https://aqueous-hollows-89178.herokuapp.com/";
+
         public TextView mName;
         public TextView mUnit;
         public TextView mTitle;
@@ -32,9 +41,24 @@ public class IncidentAdapter extends RecyclerView.Adapter<IncidentAdapter.Incide
         public Button mIncidentID;
 
 
-
         public IncidentViewHolder(@NonNull View itemView) {
             super(itemView);
+
+            HttpLoggingInterceptor loggingInterceptor =  new HttpLoggingInterceptor();
+            loggingInterceptor.setLevel(HttpLoggingInterceptor.Level.BODY);
+
+            OkHttpClient okHttpClient = new OkHttpClient.Builder()
+                    .addInterceptor(loggingInterceptor)
+                    .build();
+
+            retrofit = new Retrofit.Builder()
+                    .baseUrl(BASE_URL)
+                    .addConverterFactory(GsonConverterFactory.create())
+                    .client(okHttpClient)
+                    .build();
+
+            retrofitInterface = retrofit.create(RetrofitInterface.class);
+
             mName = itemView.findViewById(R.id.Name);
             mUnit = itemView.findViewById(R.id.Unit);
             mTitle = itemView.findViewById(R.id.IncidentTitle);
@@ -42,6 +66,30 @@ public class IncidentAdapter extends RecyclerView.Adapter<IncidentAdapter.Incide
             mRemarks = itemView.findViewById(R.id.IncidentRemarks);
             mStatus = itemView.findViewById(R.id.IncidentStatus);
             mIncidentID = itemView.findViewById(R.id.ResolveIncident);
+
+            mIncidentID.setOnClickListener(new View.OnClickListener() {
+                @Override
+                public void onClick(View v) {
+                    HashMap<String, String> incident_status = new HashMap<>();
+                    incident_status.put("id", mIncidentID.getTag().toString());
+                    incident_status.put("status", "Resolved");
+
+                    Call<IncidentList> call = retrofitInterface.incident_update(incident_status);
+                    call.enqueue(new Callback<IncidentList>() {
+                        @Override
+                        public void onResponse(Call<IncidentList> call, Response<IncidentList> response) {
+                            mIncidentID.setVisibility(View.INVISIBLE);
+                            mStatus.setText("Resolved");
+                        }
+
+                        @Override
+                        public void onFailure(Call<IncidentList> call, Throwable t) {
+//                            Toast.makeText(, "error" + t.getMessage(), Toast.LENGTH_SHORT).show();
+                        }
+                    });
+
+                }
+            });
         }
     }
 
@@ -54,6 +102,7 @@ public class IncidentAdapter extends RecyclerView.Adapter<IncidentAdapter.Incide
     public IncidentViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         View v = LayoutInflater.from(parent.getContext()).inflate(R.layout.incident_iterate, parent, false);
         IncidentViewHolder ivh = new IncidentViewHolder(v);
+
         return ivh;
 
     }
